Add promotional carousel and section header; refactor image handling and API integration
