+++ dv-marketplace/src/components/pages/MarketplaceHome/CarouselImage.tsx
@@ -1,32 +1,21 @@
-import type * as React from 'react';
+import { Carousel } from 'react-bootstrap';
 
-export interface ExampleCarouselImageProps {
+export interface PromoCarrouselItemContenProps {
   title?: string | undefined;
   text: string;
+  url: string;
 }
 
-const CarouselImage: React.FC<ExampleCarouselImageProps> = ({
-  title = 'Placeholder',
-  text,
-}) => (
-  <svg
-    className="bd-placeholder-img bd-placeholder-img-lg d-block w-100"
-    width="800"
-    height="400"
-    xmlns="http://www.w3.org/2000/svg"
-    role="img"
-    aria-label={`${title}: ${text}`}
-    preserveAspectRatio="xMidYMid slice"
-    focusable="false"
-  >
-    <title>{title}</title>
-    <rect width="100%" height="100%" fill="#555" />
-    <text x="50%" y="50%" fill="#333" dy=".3em">
-      {text}
-    </text>
-  </svg>
+const CarouselImage = ({title, text, url}:PromoCarrouselItemContenProps ) => (
+  <div>
+    <img className="card-img-top"
+            src={url}
+            alt="DCM at UNC" />
+   <Carousel.Caption>
+      <h3>{title}</h3>
+      <p>{text}</p>
+    </Carousel.Caption>
+  </div>
+
 );
-
-CarouselImage.displayName = 'ExampleCarouselImage';
-
 export default CarouselImage;
