--- dv-marketplace/src/components/pages/MarketplaceHome/Carrousel.tsx
@@ -1,37 +1,0 @@
-import Carousel from 'react-bootstrap/Carousel';
-import CarouselImage from './CarouselImage';
-
-
-
-
-function CarouselFadeExample() {
-  return (
-    <Carousel fade className='px-0'>
-      <Carousel.Item>
-        <CarouselImage text="First slide" />
-        <Carousel.Caption>
-          <h3>First slide label</h3>
-          <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>          
-        </Carousel.Caption>
-      </Carousel.Item>
-      <Carousel.Item>
-        <CarouselImage text="Second slide" />
-        <Carousel.Caption>
-          <h3>Second slide label</h3>
-          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-        </Carousel.Caption>
-      </Carousel.Item>
-      <Carousel.Item>
-        <CarouselImage text="Third slide" />
-        <Carousel.Caption>
-          <h3>Third slide label</h3>
-          <p>
-            Praesent commodo cursus magna, vel scelerisque nisl consectetur.
-          </p>
-        </Carousel.Caption>
-      </Carousel.Item>
-    </Carousel>
-  );
-}
-
-export default CarouselFadeExample;
